--- docs/guides/usage.md
@@ -1,4 +1,0 @@
----
-title: Usage
-description: Learn how to use Command in your Swift projects.
----
